Update exp and exp_horizon to be more readable
